# internal/api/wellknown/openapi.json
{
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "See docs/10-running-in-production.md for token issuance. A route with no matching scope in the token, and no admin role, is refused with 403.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Every route declared in internal/apispec, the package cmd/controller/main.go also builds its real router from, served under /api/v1. A listed route's method, pattern, scope, and link relation are the ones the server enforces. That a listed route is registered at all is not checked by anything, so confirm a route against a running controller before building on it.",
    "title": "Pleiades control plane API",
    "version": "unreleased"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs/{id}/logs": {
      "get": {
        "description": "id is validated as a UUID before use: it is concatenated into a NATS subject, and an unvalidated value could widen a wildcard subject to stream every job in the system. A fresh, ephemeral JetStream consumer scoped to exactly this job ID is created per request, so each concurrent viewer sees only the job it asked for. Each SSE frame's data is the raw JSON event payload the engine publishes per task status change, best-effort masked through every secret known at the moment that event was published (see docs/10-running-in-production.md's data handling section for the precise, non-retroactive scope of that masking).",
        "operationId": "stream_job_logs",
        "parameters": [
          {
            "description": "The job ID, a UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An open SSE stream. Never terminates on its own; the client closes it."
          },
          "400": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "id is not a UUID."
          },
          "500": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "description": "",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The JetStream consumer could not be created."
          }
        },
        "security": [
          {
            "bearerAuth": [
              "job:read"
            ]
          }
        ],
        "summary": "Stream a job's live progress over Server-Sent Events",
        "x-link-relation": "logs"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Relative to the controller's own origin.",
      "url": "/api/v1"
    }
  ]
}